+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
-
 ### Changed
+
+- Changed the default control-plane API port to 49421 and the metrics port to 49422.
+- Changed the default tunnel port to 49423.
+- Changed the default node-agent metrics port to 49431.
+- Changed the default gateway ports to 49441 (admin) and 49442 (listener).
 
 ## [0.4.0] - 2025-12-17